splitting SingleCard Component
